Initial Commit - 202302100933
Added the GUI in a relatively close to final state.

Added Visitor Class that will hold the data for the visitors that are currently in the MakerSpace and visitors that have visited for the day.

Added the Class Class to hold the data for a given course that has been imported in from the current course catalog.

Added the string array for the classes to hold the class detail read from the csv file. However, it will be later changed to hold the class class object. The data file was retrieved using a SQL query from student accounts. The SQL Query is called: NPCC_SCHED_DIV

Added the string array for the orgs used to hold all of the different orgs on campus. The orgs are pulled directly from: https://np.edu/student-services/student-organizations.aspx

Added the reasons enum that holds the different reasons a visitor would be using the space. This is hardcoded and was previously read in from a file. Since, this is unlikely to be changed, the categories were changed to being hardcoded.

Added the ReasonToString function which converts the enum to a string name version.

Added the ReadFiles function the reads the classes and orgs lists in from text files.

Added the load_cbReason that loops through the Reasons, converts to a string, then adds them to the combobox.

Added the search button function. It checks for an empty textbox, determines if the user is searching the class list or the orgs list, and adds the results to the lbClassList listbox. All while wrapped in a try statement to capture any errors.

Added functionality to change the check-in/out button to reflect in or out status.

Added the cbReason_SelectedIndexChanged function that clears the search results listbox and uses a switch statement to capture the switching of the combobox. Once, the switch is made logic is applied to determine what the forms state needs to be in for the particular Reason.

Added the WriteData function that will be later written to write the data to various files.

Added the DataValidation function that will later be sed to validate all of the data on the forms has been filled out correctly.

Added the tbSearch_KeyUp function to allow the user to enter the enter key when searching the class/org lists instead of the search button.