Commit - 202302101228
Added ability for the loading of classes to be displayed in the statusbar.

Added the class members and properties for the visitor class. Made the visitor constructor non-nullable to make sure all information will be entered.

Renamed the class Class to a non-keyword named NPC_Class along with some of the members of the class. Also, added a partial constructor to only capture the absolute necessary items and the full constructor to capture everything. Added the getters and setters for the necessary items only.

Added a list for the NPC_Class object as well as for the visitors.

Added the ability to read the delimited file and assemble the NPC_Class object using the partial constructor whilst display the progress in the statusbar.