Commit - 202302101558
Added visitor_list and visitor_in_space_list to add the visitor to the list of everyone for the day and for those that are currently in the space.

Moved classQryResult to be a global variable. Did this so that when a search is performed the original list item can be pulled up to write the details in the file instead of the curated list item.

Removed _classes as the data is now stored in the list of NPC_Classes

Added FormatListBoxEntry to format the text for easier selection by the user.

Switched the search from using findall on an array to a list for the new list of NPC_Class objects.

Completed the writeData function to output the data to a file using the id, reason for the visit, reason detail (original course description for a class or student org for a a student org reason), time and estimated hours.

Added more details to the btnComplete_Click function. Retrieved the epoch time, the reason details, and created the visitor class.

Added access to the on screen keyboard if we use a touch screen interface.