Commit - 202302150207
Added a new testing branch to test the ability to have a button click style check out.

Added a flow layout panel to the form. Also, added dynamic button generation for when a visitor enters the space a new button will be created in the flow layout panel. This allows a person to push a button to check them out via their ID. Also, added the UserCheckOut function that handles querying the visitor in the list and removing them and the button.