Commit - 202302161005
Deleted master_avatar_box branch

Added logout scheduler section which, gets the time to midnight in milliseconds and sets a timer to run every night.

Added the scheduler function to be called when the timer runs. It then calls the logout function which logs everyone out. It the calls the generate report function.