/*
Commit - 202303091203

(Designer) Added a Tool Strip Menu Button to manually invoke the end of day reporting functionality.

(Code) Added visitor direction enum to classify whether the visitor is going in or going out. Added the autologout and direction variables. The autologout is a variable to hold whether the visitor was logged out by the automatic logout function o was manually logged out. Added a default constructor, copy constructor, and a copy-modify constructor to make a copy of the visitor to be used when logging the visitor out. By creating a copy of the inbound instance, all one would have to do is change the time, direction, and, autologout variables. These actions are performed via the getnewinstance function. Removed the visitor in space list as there is no point in having duplicate data. Now, everything is handled via one visitor list. Added temp visitors (commented out) to allow for testing under certain conditions. Made several changes to allow for the controls to be returned to specific locations (setting the active control) after a task has been completed. Reformatted the output data to include the new variables defined previously in the Visitor Class. Added details to the DataValidation function to validate the data that has been received from the form and if not return a message to the user. Added details to the FormValidation function to ensure that everything is filled out the way it should be on the form. If not the user is notified and the active control property is changed to the erroneous control. On the btnCompleteClick function, I added the formValidation section to ensure the form is filled out before proceeding as well as defining the visitor direction. Added the functionality to the tool strip menu of the manual report generation function call. This allows a user to generate the report without having to wait till midnight before seeing the data. CheckOutAll and UserCheckOut functions have been modified to include the coy of the inbound visitor class object via the GetNewInstance call. The remove all avatars procedure has been modified to be used via the invoke method. This allows me to access the control buttons since they are on another thread.

completed the scheduler by starting it :)