Commit - 202303130035
Had problems with the last branch merge.

Added the function tbID_Leave that detects when the user has left the tbID textbox and moved to a different control. With this function, we perform a duplicate user check before adding the visitor via the complete button.

Changed the btnComplete_Click function to automagically identify if the tbID exists already in the system and marks them as in/out as necessary. Switched from manually adding the avatar button to calling the function as its supposed to be. Added the sections to determine if the visitor is going in or out. This is based on the visitors name being in the list of visitors for the day.

In the FormValidation function, I added a check for when the user selects the empty item in the checkbox.