Commit - 20230314 (Pi Day!)
Added settings to the app and added a bunch of functions to supplement.

Added File Validation that calls separate functions that each check if the settings exist and if they do not, then the user is asked to find the files. This includes the class, orgs, reasons, and log files.
Added an icon